# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "心愿墙公益项目API",
        "title": "心愿墙 API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/api/v1/wishes": {
            "get": {
                "description": "获取心愿列表，支持分页和过滤",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "心愿"
                ],
                "summary": "[小程序/后台]获取心愿列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "按心愿内容模糊搜索",
                        "name": "content",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "按完成状态过滤,默认为false",
                        "name": "is-done",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码，默认1",
                        "name": "page-index",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量，默认10",
                        "name": "page-size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回心愿列表和分页信息",
                        "schema": {
                            "$ref": "#/definitions/controllers.GetWishesResponse"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "创建一个新的心愿",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "心愿"
                ],
                "summary": "[后台]创建新心愿",
                "parameters": [
                    {
                        "description": "心愿信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.CreateWishRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "返回创建的心愿",
                        "schema": {
                            "$ref": "#/definitions/models.Wish"
                        }
                    },
                    "400": {
                        "description": "请求数据无效",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/wishes/{id}": {
            "put": {
                "description": "更新心愿",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "心愿"
                ],
                "summary": "[后台]更新心愿",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "心愿ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "心愿信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.UpdateWishRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回更新后的心愿",
                        "schema": {
                            "$ref": "#/definitions/models.Wish"
                        }
                    },
                    "400": {
                        "description": "请求数据无效",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "删除心愿",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "心愿"
                ],
                "summary": "[后台]删除心愿",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "心愿ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功删除心愿",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "请求数据无效",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/wishes/{id}/donor": {
            "put": {
                "description": "创建一条认领记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "心愿"
                ],
                "summary": "[小程序]点亮心愿",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "心愿ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "捐赠者信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.UpdateWishDonorRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回更新后的心愿",
                        "schema": {
                            "$ref": "#/definitions/models.Wish"
                        }
                    },
                    "400": {
                        "description": "请求数据无效",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "心愿不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "controllers.CreateWishRequest": {
            "type": "object",
            "properties": {
                "childName": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "gender": {
                    "$ref": "#/definitions/models.Gender"
                },
                "grade": {
                    "type": "string"
                },
                "isPublished": {
                    "type": "boolean"
                },
                "photoUrl": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                }
            }
        },
...
        "controllers.GetWishesResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Wish"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/utils.Pagination"
                }
            }
        },
...
        "controllers.UpdateWishDonorRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "comment": {
                    "type": "string"
                },
                "donorMobile": {
                    "type": "string"
                },
                "donorName": {
                    "type": "string"
                }
            }
        },
        "controllers.UpdateWishRequest": {
            "type": "object",
            "properties": {
                "childName": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "gender": {
                    "$ref": "#/definitions/models.Gender"
                },
                "grade": {
                    "type": "string"
                },
                "isPublished": {
                    "type": "boolean"
                },
                "photoUrl": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                }
            }
        },
...
        "models.Gender": {
            "description": "用户性别类型",
            "type": "string",
            "enum": [
                "male",
                "female"
            ],
            "x-enum-varnames": [
                "Male",
                "Female"
            ]
        },
        "models.User": {
            "description": "微信小程序用户信息",
            "type": "object",
            "properties": {
                "avatarUrl": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "integer"
                },
                "deletedAt": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "integer"
                },
                "wechatOpenId": {
                    "type": "string"
                },
                "wechatUnionId": {
                    "type": "string"
                }
            }
        },
        "models.Wish": {
            "description": "心愿信息",
            "type": "object",
            "properties": {
                "activeRecord": {
                    "$ref": "#/definitions/models.WishRecord"
                },
                "activeRecordId": {
                    "type": "integer"
                },
                "childName": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "integer"
                },
                "deletedAt": {
                    "type": "integer"
                },
                "gender": {
                    "$ref": "#/definitions/models.Gender"
                },
                "grade": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isPublished": {
                    "type": "boolean"
                },
                "photoUrl": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "integer"
                }
            }
        },
        "models.WishRecord": {
            "description": "心愿认领记录",
            "type": "object",
            "properties": {
                "cancellationTime": {
                    "description": "取消时间",
                    "type": "integer"
                },
                "confirmationMessage": {
                    "description": "确认信息",
                    "type": "string"
                },
                "confirmationPhotos": {
                    "description": "确认照片数组",
                    "type": "string"
                },
                "confirmationTime": {
                    "description": "确认时间",
                    "type": "integer"
                },
                "createdAt": {
                    "type": "integer"
                },
                "deletedAt": {
                    "type": "integer"
                },
                "deliveryNumber": {
                    "description": "发货单号",
                    "type": "string"
                },
                "deliveryTime": {
                    "description": "发货时间",
                    "type": "integer"
                },
                "donor": {
                    "$ref": "#/definitions/models.User"
                },
                "donorAddress": {
                    "type": "string"
                },
                "donorComment": {
                    "type": "string"
                },
                "donorId": {
                    "type": "integer"
                },
                "donorMobile": {
                    "type": "string"
                },
                "donorName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "ownerGiftMessage": {
                    "description": "心愿主人回礼信息",
                    "type": "string"
                },
                "ownerGiftPhotos": {
                    "description": "心愿主人回礼照片数组",
                    "type": "string"
                },
                "ownerGiftTime": {
                    "description": "心愿主人回礼时间",
                    "type": "integer"
                },
                "platformGiftMessage": {
                    "description": "平台回礼信息",
                    "type": "string"
                },
                "platformGiftPhotos": {
                    "description": "平台回礼照片数组",
                    "type": "string"
                },
                "platformGiftTime": {
                    "description": "平台回礼时间",
                    "type": "integer"
                },
                "receiptMessage": {
                    "description": "签收信息",
                    "type": "string"
                },
                "receiptPhotos": {
                    "description": "签收照片数组",
                    "type": "string"
                },
                "receiptTime": {
                    "description": "签收时间",
                    "type": "integer"
                },
                "shippingNumber": {
                    "description": "寄送单号",
                    "type": "string"
                },
                "shippingTime": {
                    "description": "寄送时间",
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/models.WishRecordStatus"
                },
                "updatedAt": {
                    "type": "integer"
                },
                "wish": {
                    "$ref": "#/definitions/models.Wish"
                },
                "wishId": {
                    "type": "integer"
                }
            }
        },
        "models.WishRecordStatus": {
            "description": "心愿认领记录状态",
            "type": "string",
            "enum": [
                "pending_shipment",
                "pending_confirmation",
                "confirmed",
                "awaiting_receipt",
                "completed",
                "gift_returned",
                "cancelled"
            ],
            "x-enum-varnames": [
                "StatusPendingShipment",
                "StatusPendingConfirmation",
                "StatusConfirmed",
                "StatusAwaitingReceipt",
                "StatusCompleted",
                "StatusGiftReturned",
                "StatusCancelled"
            ]
        },
        "utils.Pagination": {
            "type": "object",
            "properties": {
                "pageIndex": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "pageTotal": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        }
    }
}